WIP
No major changes, just reshuffling, will come back and work on this soon. (stuff might be broken here)
