+++ constants.py
@@ -1,0 +1,8 @@
+prefixes = {1: "meth", 2: "eth", 3: "prop", 4: "but", 5: "pent", 6: "hex", 7: "hept", 8: "oct", 9: "non", 10: "dec",
+            11: "undec", 12: "dodec", 13: "tridec", 14: "tetradec", 15: "pentadec", 16: "hexadec", 17: "heptadec",
+            18: "octadec", 19: "nonadec", 20: "icos"}
+
+# precedence = {"=": 1, "~": 1}
+
+multipl_suffixes = {2: "di", 3: "tri", 4: "tetra", 5: "penta", 6: "hexa", 7: "hepta", 8: "octa", 9: "nona"}
+symbol = '\u2261'  # The triple bond symbol ≡
