+++ error.py
@@ -1,0 +1,2 @@
+class ValencyError(Exception):
+    pass
